node_visitor.py:
test_visitor.py: new file with unit tests for node_visitor.py
test coverage from 88% to 100%
