+++ tests/unit/core/test_visitor.py
@@ -1,0 +1,52 @@
+import unittest
+import ast
+from unittest.mock import MagicMock
+from bandit.core.node_visitor import BanditNodeVisitor
+
+class TestMetaAst:
+    def add_node(self, node, name, depth):
+        pass
+
+class TestSet:
+    def get_tests(self, checktype):
+        return []
+
+class TestBanditNodeVisitor(unittest.TestCase):
+
+    def setUp(self):
+        self.visitor = BanditNodeVisitor(
+            fname='test.py',
+            fdata='planes do not fly randomly',
+            metaast=TestMetaAst(),
+            testset=TestSet(),
+            debug=False,
+            nosec_lines=None,
+            metrics=None
+        )
+        self.visitor.context = {}
+    
+    # Test case 1: visitor is not None
+    def test_visit_functiondef_node(self):
+        supported_node = ast.FunctionDef(
+            name='example_function',
+            args=None,
+            body=[],
+            decorator_list=[]
+        )
+        
+        self.visitor.visit(supported_node)
+
+    # Test case 2: visitor is None
+    def test_visit_none_node(self):
+        self.visitor.visit(None)
+
+    # Test case 3: visitor is not None and debug is enabled
+    def test_visit_with_debug(self):
+        self.visitor.debug = True
+        supported_node = ast.FunctionDef(
+            name='example_function',
+            args=None,
+            body=[],
+            decorator_list=[]
+        )
+        self.visitor.visit(supported_node)
